requirements.txt
glob2
opencv-python
numpy
tensorflow
scikit-learn
keras
matplotlib
flask
google-cloud-storage
gunicorn
firebase-admin
requests
pymysql